+++ projects/angular-cesium/src/lib/angular-cesium-widgets/components.ts
@@ -1,4 +1,5 @@
 export * from './components/polygons-editor/polygons-editor.component';
+export * from './components/rectangles-editor/rectangles-editor.component';
 export * from './components/circles-editor/circles-editor.component';
 export * from './components/polylines-editor/polylines-editor.component';
 export * from './components/hippodrome-editor/hippodrome-editor.component';