+++ projects/angular-cesium/src/lib/angular-cesium/models/rectangle.ts
@@ -1,0 +1,12 @@
+export interface Rectangle {
+  west: number;
+  south: number;
+  east: number;
+  north: number;
+  width?: number;
+
+  equals(right: Rectangle): boolean;
+
+  clone(): Rectangle;
+}
+